+++ app/src/main/java/com/rama/teyin/activities/MainActivity.kt
@@ -41,7 +41,7 @@ import com.rama.teyin.adapters.DirectoryListAdapter
 import com.rama.teyin.adapters.FileListAdapter
 import com.rama.teyin.managers.FileManager
 import com.rama.teyin.managers.PrefsManager as BohioPrefsManager
-import com.rama.teyin.managers.ThemeManager
+import com.rama.bohio.managers.ThemeManager
 import java.io.File
 import java.text.SimpleDateFormat
 import java.util.Date
@@ -163,7 +163,7 @@ class MainActivity : CsActivity() {
         directoriesButton.setOnClickListener {
             showDirs = !showDirs
             if (showDirs) {
-                it.setBackgroundColor(resources.getColor(BohioR.color.button_selected))
+                it.setBackgroundColor(resources.getColor(BohioR.color.bg_2))
                 directoriesFragment.visibility = View.VISIBLE
                 filesFragment.visibility = View.GONE
                 refreshFavorites()